requirements.txt
flask
mysql.connector
flask_cors
flask_restful
flask_jsonpify